# Conflicts:
#	.github/workflows/build.yml
#	.github/workflows/pr.yml

+++ makeimage.sh
@@ -18,13 +18,12 @@ docker buildx inspect ffbuilder &>/dev/null || docker buildx create \
     --driver-opt env.BUILDKIT_STEP_LOG_MAX_SIZE=-1 \
     --driver-opt env.BUILDKIT_STEP_LOG_MAX_SPEED=-1
 
-if [[ -z "$QUICKBUILD" ]]; then
-    docker container inspect ffbuildreg &>/dev/null || \
-        docker run --rm -d -p 127.0.0.1:0:5000 --name ffbuildreg registry:2
-    LOCAL_REG_PORT="$(docker container inspect --format='{{range $p, $conf := .NetworkSettings.Ports}}{{(index $conf 0).HostPort}}{{end}}' ffbuildreg)"
-    LOCAL_ROOT="127.0.0.1:${LOCAL_REG_PORT}/local"
-    trap "rm -f '$TMPCFG'; docker container stop ffbuildreg" EXIT
+docker container inspect ffbuildreg &>/dev/null || \
+    docker run --rm -d -p 127.0.0.1:64647:5000 --name ffbuildreg registry:2
+LOCAL_REG_PORT="$(docker container inspect --format='{{range $p, $conf := .NetworkSettings.Ports}}{{(index $conf 0).HostPort}}{{end}}' ffbuildreg)"
+LOCAL_ROOT="127.0.0.1:${LOCAL_REG_PORT}/local"
 
+if [[ -z "$QUICKBUILD" ]]; then
     if grep "FROM.*base.*" "images/base-${TARGET}/Dockerfile" >/dev/null 2>&1; then
         docker buildx --builder ffbuilder build \
             --cache-from=type=local,src=.cache/"${BASE_IMAGE/:/_}" \
@@ -41,6 +40,7 @@ if [[ -z "$QUICKBUILD" ]]; then
     export REGISTRY_OVERRIDE="127.0.0.1:${LOCAL_REG_PORT}" GITHUB_REPOSITORY="local"
 fi
 
+./download.sh
 ./generate.sh "$TARGET" "$VARIANT" "${ADDINS[@]}"
 
 docker buildx --builder ffbuilder build \
@@ -48,4 +48,7 @@ docker buildx --builder ffbuilder build \
     --cache-to=type=local,mode=max,dest=.cache/"${IMAGE/:/_}" \
     --load --tag "$IMAGE" .
 
-docker buildx rm -f ffbuilder
+if [[ -z "$NOCLEAN" ]]; then
+    docker container stop ffbuildreg
+    docker buildx rm -f ffbuilder
+fi
